+++ DataStructures/TurnInfoFactory.h
@@ -102,15 +102,15 @@ public:
             }
             //merge edges (s,t) and (t,s) into bidirectional edge
             if ( forwardEdge.data.distance == backwardEdge.data.distance ) {
-                if ( forwardEdge.data.distance != std::numeric_limits< int >::max() ) {
+                if ( (int)forwardEdge.data.distance != std::numeric_limits< int >::max() ) {
                     forwardEdge.data.backward = true;
                     edges[edge++] = forwardEdge;
                 }
             } else { //insert seperate edges
-                if ( forwardEdge.data.distance != std::numeric_limits< int >::max() ) {
+                if ( (int)forwardEdge.data.distance != std::numeric_limits< int >::max() ) {
                     edges[edge++] = forwardEdge;
                 }
-                if ( backwardEdge.data.distance != std::numeric_limits< int >::max() ) {
+                if ( (int)backwardEdge.data.distance != std::numeric_limits< int >::max() ) {
                     edges[edge++] = backwardEdge;
                 }
             }
